[testUtils] Improve output when expected success is a failure
When we are expecting the result to be successful but it isn't, we were only
getting output that success != true, which doesn't tell you much - it's better
to see the errors (that shouldn't be happening).

In this case (expecting success but actually it's a failure), we can assert that
the errors should be empty and that will cause the errors to print out first,
which is more useful for debugging.