[CI] Add GitHub Actions CI Config (#6)
* [Build] Add 'yarn lint' command

Add a basic 'yarn lint' command in all packages, and in the root package.

* Fix dependencies in nest package

- Add express as dev dependency to satisfy linter
- Relax rxjs version to satisfy yarn warning when we use this in our NestJs app.

* [Jest] Add jest-env.ts to tsconfig

These tests were working for us locally but failing with type errors in
CI - adding the file to tsconfig makes it work.

* [CI] Add GitHub Actions config

Add a basic CI config that runs lint, build, and tests in parallel jobs.

Includes 'slack-ci' script for easily sending slack notifications to our
internal Slack. Usage and inline help in the script.