Remove notBlankString (#19)
notBlankString is not an appropriate name as the codec also trims otherwise valid input, which may be surprising to someone only expecting that it would validate against blanks. It was intended as a factory helper for people accepting string inputs that they wanted to trim and not accept empty strings. Consumers can still create easily create that codec in their own project. If demand arises this can be later introduced as an export by Salus.